Create HoverBoardDR.ino
just a mash-up of the motor control code and the audio processing code. This is a rough idea of what it could look like.

just a side note but I think we might want to add a counter that counts the amount of time a certain frequency is detected. Doing this could help prevent the accidental triggering of the code due to noise.  Basically, we have  to give the microphone a certain frequency for a short time before it executes the command associated with the frequency